Revert "Fixes Bug 1205350 - collector2015 - allow multiple collector types"
