+++ socorro/collector/collector_app.py
@@ -11,8 +11,7 @@
 # set both socorro and configman in your PYTHONPATH
 
 from socorro.app.generic_app import App, main
-from socorro.webapi.classPartial import class_with_partial_init
-from socorro.lib.converters import web_services_from_str
+
 
 from configman import Namespace
 from configman.converters import class_converter
@@ -24,30 +23,7 @@ application = None
 
 
 #==============================================================================
-class BaseCollectorApp(App):
-    app_name = 'collector'
-    app_version = '5.0'
-    app_description = __doc__
-
-    #--------------------------------------------------------------------------
-    # in this section, define any configuration requirements
-    required_config = Namespace()
-
-    #--------------------------------------------------------------------------
-    # web_server namespace
-    #     the namespace is for config parameters the web server
-    #--------------------------------------------------------------------------
-    required_config.namespace('web_server')
-    required_config.web_server.add_option(
-        'wsgi_server_class',
-        doc='a class implementing a wsgi web server',
-        default='socorro.webapi.servers.CherryPy',
-        from_string_converter=class_converter
-    )
-
-
-#==============================================================================
-class CollectorApp(BaseCollectorApp):
+class CollectorApp(App):
     app_name = 'collector'
     app_version = '4.0'
     app_description = __doc__
@@ -94,8 +70,20 @@ class CollectorApp(BaseCollectorApp):
     )
 
     #--------------------------------------------------------------------------
+    # web_server namespace
+    #     the namespace is for config parameters the web server
+    #--------------------------------------------------------------------------
+    required_config.namespace('web_server')
+    required_config.web_server.add_option(
+        'wsgi_server_class',
+        doc='a class implementing a wsgi web server',
+        default='socorro.webapi.servers.CherryPy',
+        from_string_converter=class_converter
+    )
+
+    #--------------------------------------------------------------------------
     def main(self):
-        # modwsgi requires a module level name 'application'
+        # modwsgi requireds a module level name 'application'
         global application
 
         services_list = (
@@ -112,78 +100,6 @@ class CollectorApp(BaseCollectorApp):
             services_list
         )
 
-        # for modwsgi the 'run' method returns the wsgi function that the web
-        # server will use.  For other webservers, the 'run' method actually
-        # starts the standalone web server.
-        application = self.web_server.run()
-
-
-#==============================================================================
-class Collector2015App(BaseCollectorApp):
-    #--------------------------------------------------------------------------
-    # in this section, define any configuration requirements
-    required_config = Namespace()
-
-    #--------------------------------------------------------------------------
-    # services namespace
-    #     the namespace is for config parameters about how to interpret
-    #     crash submissions
-    #--------------------------------------------------------------------------
-    required_config.namespace('services')
-    required_config.services.add_option(
-        'services_controller',
-        default='''[
-        {
-            "name": "collector",
-            "uri": "/submit",
-            "service_implementation_class":
-            "socorro.collector.wsgi_breakpad_collector.BreakpadCollector2015"
-        },
-        {
-            "name": "generic",
-            "uri": "/some/other/uri",
-            "service_implementation_class":
-                "socorro.collector.generic_collector.GenericCollector"
-        }
-        ]''',
-        doc='json-like list of services to be offered by this collector:'
-            '[{"name": sevice-name, "uri": uri, '
-            '"service_implementation_class": class-name}, ...]',
-        from_string_converter=web_services_from_str(),
-    )
-
-    #--------------------------------------------------------------------------
-    def main(self):
-        # modwsgi requires a module level name 'application'
-        global application
-
-        services_list = []
-        # take the list of services that have been specified in the configman
-        # setup and create a list of services appropriate for the wsgi server
-        # class.  Ensure that each service has been coupled with its namespace
-        # from the final configuration
-        for namespace, uri, service_class in (
-            self.config.services.services_controller.service_list
-        ):
-            services_list.append(
-                # a tuple associating a URI with a service class
-                (
-                    uri,
-                    # a binding of the service class with the configuration
-                    # namespace for that service class
-                    class_with_partial_init(
-                        self.config.services[namespace]
-                            .service_implementation_class,
-                        self.config.services[namespace]
-                    )
-                )
-            )
-
-        # initialize the wsgi server with the list of URI/service impl tuples
-        self.web_server = self.config.web_server.wsgi_server_class(
-            self.config,  # needs the whole config not the local namespace
-            services_list
-        )
 
         # for modwsgi the 'run' method returns the wsgi function that the web
         # server will use.  For other webservers, the 'run' method actually
