+++ socorro/collector/wsgi_breakpad_collector.py
@@ -2,18 +2,24 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
+import web
 import time
+import zlib
+import cgi
+import cStringIO
+
+from contextlib import closing
 
 from socorro.lib.ooid import createNewOoid
+from socorro.lib.util import DotDict
 from socorro.collector.throttler import DISCARD, IGNORE
 from socorro.lib.datetimeutil import utc_now
-from socorro.collector.wsgi_generic_collector import GenericCollectorBase
 
-from configman import Namespace, class_converter
+from configman import RequiredConfig, Namespace, class_converter
 
 
 #==============================================================================
-class BreakpadCollectorBase(GenericCollectorBase):
+class BreakpadCollector(RequiredConfig):
     #--------------------------------------------------------------------------
     # in this section, define any configuration requirements
     required_config = Namespace()
@@ -28,41 +34,99 @@ class BreakpadCollectorBase(GenericCollectorBase):
         default='bp-'
     )
     required_config.add_option(
+        'accept_submitted_crash_id',
+        doc='a boolean telling the collector to use a crash_id provided in '
+            'the crash submission',
+        default=False
+    )
+    required_config.add_option(
         'accept_submitted_legacy_processing',
         doc='a boolean telling the collector to use a any legacy_processing'
             'flag submitted with the crash',
         default=False
     )
+    required_config.add_option(
+        'checksum_method',
+        doc='a reference to method that accepts a string and calculates a'
+            'hash value',
+        default='hashlib.md5',
+        from_string_converter=class_converter
+    )
 
     #--------------------------------------------------------------------------
     def __init__(self, config):
-        super(BreakpadCollectorBase, self).__init__(config)
-        self.dump_field = self._get_dump_field()
-        self.dump_id_prefix = self._get_dump_id_prefix()
-        self.accept_submitted_legacy_processing = \
-            self._get_accept_submitted_legacy_processing()
-        self.throttler = self._get_throttler()
-        self.crash_storage = self._get_crash_storage()
+        self.config = config
+        self.logger = self.config.logger
+        self.throttler = config.throttler
+        self.dump_id_prefix = config.collector.dump_id_prefix
+        self.crash_storage = config.crash_storage
+        self.dump_field = config.collector.dump_field
 
     #--------------------------------------------------------------------------
-    def _get_dump_field(self):
-        return self.config.dump_field
+    uri = '/submit'
+    #--------------------------------------------------------------------------
 
     #--------------------------------------------------------------------------
-    def _get_dump_id_prefix(self):
-        return self.config.dump_id_prefix
+    @staticmethod
+    def _process_fieldstorage(fs):
+        if isinstance(fs, list):
+            return [_process_fieldstorage(x) for x in fs]
+        elif fs.filename is None:
+            return fs.value
+        else:
+            return fs
 
     #--------------------------------------------------------------------------
-    def _get_accept_submitted_legacy_processing(self):
-        return self.config.accept_submitted_legacy_processing
+    def _form_as_mapping(self):
+        """this method returns the POST form mapping with any gzip
+        decompression automatically handled"""
+        if web.ctx.env.get('HTTP_CONTENT_ENCODING') == 'gzip':
+            # Handle gzipped form posts
+            gzip_header = 16 + zlib.MAX_WBITS
+            data = zlib.decompress(web.webapi.data(), gzip_header)
+            e = web.ctx.env.copy()
+            with closing(cStringIO.StringIO(data)) as fp:
+                # this is how web.webapi.rawinput() handles
+                # multipart/form-data, as of this writing
+                fs = cgi.FieldStorage(fp=fp, environ=e, keep_blank_values=1)
+                form = web.utils.storage(
+                    [(k, self._process_fieldstorage(fs[k])) for k in fs.keys()]
+                )
+                return form
+        return web.webapi.rawinput()
 
     #--------------------------------------------------------------------------
-    def _get_throttler(self):
-        return self.config.throttler
+    @staticmethod
+    def _no_x00_character(value):
+        if isinstance(value, unicode) and u'\u0000' in value:
+            return ''.join(c for c in value if c != u'\u0000')
+        if isinstance(value, str) and '\x00' in value:
+            return ''.join(c for c in value if c != '\x00')
+        return value
 
     #--------------------------------------------------------------------------
-    def _get_crash_storage(self):
-        return self.config.crash_storage
+    def _get_raw_crash_from_form(self):
+        """this method creates the raw_crash and the dumps mapping using the
+        POST form"""
+        dumps = DotDict()
+        raw_crash = DotDict()
+        raw_crash.dump_checksums = DotDict()
+        for name, value in self._form_as_mapping().iteritems():
+            name = self._no_x00_character(name)
+            if isinstance(value, basestring):
+                if name != "dump_checksums":
+                    raw_crash[name] = self._no_x00_character(value)
+            elif hasattr(value, 'file') and hasattr(value, 'value'):
+                dumps[name] = value.value
+                raw_crash.dump_checksums[name] = \
+                    self.config.collector.checksum_method(
+                        value.value
+                    ).hexdigest()
+            elif isinstance(value, int):
+                raw_crash[name] = value
+            else:
+                raw_crash[name] = value.value
+        return raw_crash, dumps
 
     #--------------------------------------------------------------------------
     def POST(self, *args):
@@ -73,7 +137,9 @@ class BreakpadCollectorBase(GenericCollectorBase):
         # legacy - ought to be removed someday
         raw_crash.timestamp = time.time()
 
-        if (not self.accept_submitted_crash_id or 'uuid' not in raw_crash):
+        if (not self.config.collector.accept_submitted_crash_id
+            or 'uuid' not in raw_crash
+        ):
             crash_id = createNewOoid(current_timestamp)
             raw_crash.uuid = crash_id
             self.logger.info('%s received', crash_id)
@@ -82,7 +148,7 @@ class BreakpadCollectorBase(GenericCollectorBase):
             self.logger.info('%s received with existing crash_id:', crash_id)
 
         if ('legacy_processing' not in raw_crash
-            or not self.accept_submitted_legacy_processing
+            or not self.config.collector.accept_submitted_legacy_processing
         ):
             raw_crash.legacy_processing, raw_crash.throttle_rate = (
                 self.throttler.throttle(raw_crash)
@@ -97,86 +163,10 @@ class BreakpadCollectorBase(GenericCollectorBase):
             self.logger.info('%s ignored', crash_id)
             return "Unsupported=1\n"
 
-        raw_crash.type_tag = self.dump_id_prefix.strip('-')
-
-        self.crash_storage.save_raw_crash(
-            raw_crash,
-            dumps,
-            crash_id
+        self.config.crash_storage.save_raw_crash(
+          raw_crash,
+          dumps,
+          crash_id
         )
         self.logger.info('%s accepted', crash_id)
         return "CrashID=%s%s\n" % (self.dump_id_prefix, crash_id)
-
-
-#==============================================================================
-class BreakpadCollector(BreakpadCollectorBase):
-    #--------------------------------------------------------------------------
-    # in this section, define any configuration requirements
-    required_config = Namespace()
-
-    #--------------------------------------------------------------------------
-    uri = '/submit'
-    #--------------------------------------------------------------------------
-
-    #--------------------------------------------------------------------------
-    def _get_dump_field(self):
-        return self.config.collector.dump_field
-
-    #--------------------------------------------------------------------------
-    def _get_dump_id_prefix(self):
-        return self.config.collector.dump_id_prefix
-
-    #--------------------------------------------------------------------------
-    def _get_accept_submitted_legacy_processing(self):
-        return self.config.collector.accept_submitted_legacy_processing
-
-    #--------------------------------------------------------------------------
-    def _get_checksum_method(self):
-        return self.config.collector.checksum_method
-
-    #--------------------------------------------------------------------------
-    def _get_accept_submitted_crash_id(self):
-        return self.config.collector.accept_submitted_crash_id
-
-
-#==============================================================================
-class BreakpadCollector2015(BreakpadCollectorBase):
-    #--------------------------------------------------------------------------
-    # in this section, define any configuration requirements
-    required_config = Namespace()
-    #--------------------------------------------------------------------------
-    # throttler namespace
-    #     the namespace is for config parameters for the throttler system
-    #--------------------------------------------------------------------------
-    required_config.namespace('throttler')
-    required_config.throttler.add_option(
-        'throttler_class',
-        default='socorro.collector.throttler.LegacyThrottler',
-        doc='the class that implements the throttling action',
-        from_string_converter=class_converter
-    )
-    #--------------------------------------------------------------------------
-    # storage namespace
-    #     the namespace is for config parameters crash storage
-    #--------------------------------------------------------------------------
-    required_config.namespace('storage')
-    required_config.storage.add_option(
-        'crashstorage_class',
-        doc='the source storage class',
-        default='socorro.external.fs.crashstorage'
-                '.FSLegacyDatedRadixTreeStorage',
-        from_string_converter=class_converter
-    )
-
-    #--------------------------------------------------------------------------
-    def _get_throttler(self):
-        return self.config.throttler.throttler_class(
-            self.config.throttler
-        )
-
-    #--------------------------------------------------------------------------
-    def _get_crash_storage(self):
-        return self.config.storage.crashstorage_class(
-            self.config.storage
-        )
-
